SAK-2838, SAK-2844 Export floating assessment scores to Gradebook instead of truncating to an integer
git-svn-id: https://source.sakaiproject.org/svn/trunk/sakai/sam@3970 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ samigo/component/src/java/org/sakaiproject/tool/assessment/integration/helper/integrated/GradebookServiceHelperImpl.java
@@ -145,7 +145,7 @@ public void removeExternalAssessment(String gradebookUId,
                               publishedAssessment.getPublishedAssessmentId().
                               toString(), null,
                               publishedAssessment.getTitle(),
-                              publishedAssessment.getTotalScore().longValue(),
+                              publishedAssessment.getTotalScore().doubleValue(),
                               publishedAssessment.getAssessmentAccessControl().
                               getDueDate(),
                               appName); // Use the app name from sakai
